requirements.txt
absl-py==2.0.0
chex==0.1.7
click==8.1.7
clu==0.0.10
contextlib2==21.6.0
dm-tree==0.1.8
etils==1.5.0
filelock==3.12.4
flax==0.7.4
fsspec==2023.9.2
importlib-resources==6.1.0
jax==0.4.18
jaxlib==0.4.18
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.1
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.1
numpy==1.23.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.1
packaging==23.2
protobuf==4.24.4
Pygments==2.16.1
PyYAML==6.0.1
rich==13.6.0
scipy==1.11.3
six==1.16.0
sympy==1.12
tensorstore==0.1.45
toolz==0.12.0
torch==2.1.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
wrapt==1.15.0
zipp==3.17.0
